Extract session-selection from auth component (#874)
Extract session-selection from auth component

Currently, the AuthComponent manages the form used
for sesssion selection and triggers the OAuth process.

Let's move this logic (and related view) into a child
component.
