+++ src/app/auth/session-selection/session-selection.component.css
@@ -1,0 +1,10 @@
+.login-field {
+  width: 100%;
+}
+
+.sign-in-button {
+  background: #f7fcfe;
+  line-height: 45px;
+  border: 1px solid currentColor;
+  width: 100%;
+}
