+++ src/app/auth/session-selection/session-selection.component.html
@@ -1,0 +1,34 @@
+<mat-card class="login-card">
+  <mat-card-header style="margin-bottom: 20px">
+    <mat-card-title class="login-title"> Select Your Session </mat-card-title>
+  </mat-card-header>
+
+  <form [formGroup]="profileForm" (ngSubmit)="setupSession()">
+    <app-profiles
+      class="login-field"
+      [urlEncodedSessionName]="this.urlEncodedSessionName"
+      (selectedProfileEmitter)="onProfileSelect($event)"
+    ></app-profiles>
+    <mat-card-content>
+      <mat-form-field class="login-field">
+        <input matInput placeholder="Settings Location (Org/Repo)" formControlName="session" required />
+      </mat-form-field>
+
+      <mat-card-actions>
+        <button
+          *ngIf="!isSettingUpSession"
+          class="sign-in-button"
+          type="submit"
+          [disabled]="profileForm.invalid || isSettingUpSession"
+          mat-stroked-button
+          color="primary"
+        >
+          <div>Submit</div>
+        </button>
+        <div *ngIf="isSettingUpSession" class="spinner-centralize">
+          <mat-spinner [diameter]="47"></mat-spinner>
+        </div>
+      </mat-card-actions>
+    </mat-card-content>
+  </form>
+</mat-card>
